refactor(services): archive Flask web_api + rename utils/unified_pipeline (#208-R)
Part 1 — utils/unified_pipeline.py → utils/analysis_config.py:
  Renamed to avoid collision with orchestration/unified_pipeline.py (3900+ lines).
  Old path is now a deprecation shim that re-exports all symbols.

Part 2 — services/web_api/ Flask layer archived:
  Flask app, routes, ASGI wrapper, start script archived to
  docs/archives/services_web_api_flask/. Business logic services
  (AnalysisService, ValidationService, FallacyService, etc.) and
  data models remain in-place — MCP server and other consumers
  continue to work via unchanged import paths.

  Superseded by FastAPI app at api/main.py (25+ routes, WebSocket,
  CapabilityRegistry + WorkflowDSL integration).

Closes #217
