+++ argumentation_analysis/services/web_api/__init__.py
@@ -1,12 +1,13 @@
 """
-Package pour les services web_api au sein de argumentation_analysis.services.
+Package web_api — business logic services and models.
+
+NOTE (2026-03-24, #217): The Flask app layer (app.py, routes/, start_api.py,
+asgi.py) has been archived to docs/archives/services_web_api_flask/.
+The modern REST API is at api/main.py (FastAPI, 25+ routes, WebSocket).
+
+The business logic services (services/*.py) and data models (models/*.py)
+remain here and are actively used by:
+  - argumentation_analysis.services.mcp_server (MCP tool exposure)
+  - argumentation_analysis.core.argumentation_analyzer
+  - Various integration tests
 """
-
-# Ce fichier __init__.py rend ce répertoire un package Python.
-# Il peut être vide ou contenir du code d'initialisation spécifique au package.
-
-# Exemple d'initialisation (optionnel):
-# print(f"Initialisation du package: {__name__}")
-
-# Vous pouvez également définir __all__ pour contrôler ce qui est importé avec 'from . import *'
-# __all__ = ['un_module_specifique']
